refactor : KuroomMap 컴포넌트 리팩토링. 로직들 KuroomMapUtils로 분리
