+++ package.json
@@ -24,6 +24,7 @@
   "devDependencies": {
     "@eslint/js": "^9.19.0",
     "@types/axios": "^0.9.36",
+    "@types/navermaps": "^3.7.9",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-react": "^4.3.4",
