+++ src/components/Map/KuroomMap.tsx
@@ -1,5 +1,9 @@
 import { useEffect, useRef, useState } from "react";
-import myMarkerIcon from "../../assets/map/mylocationMarker.svg";
+import {
+  myLocationTracking,
+  noTracking,
+  renderMarkers,
+} from "./kuroomMapUtils";
 
 interface MarkerData {
   lat: number;
@@ -29,8 +33,8 @@ const KuroomMap = ({
   zoomable = true,
 }: MapProps) => {
   const mapRef = useRef(null);
-  const markerRef = useRef<any>(null);
-  const mapInstance = useRef<any>(null); // 지도 객체를 저장할 ref
+  const markerRef = useRef<naver.maps.Marker | null>(null);
+  const mapInstance = useRef<naver.maps.Map | null>(null); // 지도 객체를 저장할 ref
   const [currentLatLng, setCurrentLatLng] = useState<any>(null); // 현재 위치를 기억
   const isTrackingRef = useRef(true); // 추적 상태 최신값을 유지할 ref
 
@@ -64,97 +68,29 @@ const KuroomMap = ({
     const map = new window.naver.maps.Map(mapRef.current, mapOptions);
     mapInstance.current = map; // 지도 인스턴스를 ref에 저장
 
-    // 마커 렌더링
-    markers.forEach(({ lat, lng, title, onClick }) => {
-      const marker = new window.naver.maps.Marker({
-        position: new window.naver.maps.LatLng(lat, lng),
-        map,
-        title,
-      });
-
-      if (onClick) {
-        window.naver.maps.Event.addListener(marker, "click", onClick);
-      }
-    });
-
-    if (setIsTracking) {
-      const disableTracking = () => {
-        setIsTracking(false);
-        isTrackingRef.current = false;
-      };
-
-      // 드래그
-      window.naver.maps.Event.addListener(map, "drag", disableTracking);
-
-      // 줌 변경 (줌 버튼, 휠, 핀치 줌 포함)
-      window.naver.maps.Event.addListener(map, "zoom_changed", disableTracking);
-
-      // 더블 클릭 줌
-      window.naver.maps.Event.addListener(map, "dblclick", disableTracking);
-
-      // 마우스 휠
-      window.naver.maps.Event.addListener(map, "wheel", disableTracking);
-    }
+    // 화면 조작 시 위치 추적 비활성화
+    if (setIsTracking) noTracking(map, setIsTracking, isTrackingRef);
 
     // 현재 위치 정보 가져와서 마커 추가 및 watchPosition으로 따라가기
     if (navigator.geolocation) {
-      const watchId = navigator.geolocation.watchPosition(
-        (position) => {
-          const { latitude, longitude } = position.coords;
-          const currentLocation = new window.naver.maps.LatLng(
-            latitude,
-            longitude
-          );
-
-          setCurrentLatLng(currentLocation); // 현재 위치 상태 업데이트
-
-          if (markerRef.current) {
-            // 기존 마커 위치 업데이트
-            markerRef.current.setPosition(currentLocation);
-          } else {
-            // 최초 마커 생성
-            markerRef.current = new window.naver.maps.Marker({
-              position: currentLocation,
-              map,
-              title: "내 위치",
-              icon: {
-                // 마커 아이콘 추가
-                url: myMarkerIcon,
-              },
-            });
-          }
-
-          if (isTrackingRef.current) {
-            // 지도 중심을 내 위치를 기준으로 이동
-            map.setCenter(currentLocation);
-          }
-        },
-        (error) => {
-          console.error("위치 정보를 가져올 수 없습니다:", error);
-
-          if (error.code === 1) {
-            alert("위치 권한이 거부되었습니다.");
-          } else if (error.code === 2) {
-            alert("위치 정보를 사용할 수 없습니다.");
-          } else if (error.code === 3) {
-            alert("위치 정보를 가져오는 데 시간이 초과되었습니다.");
-          }
-        },
-        {
-          enableHighAccuracy: true, // 정확도 향상
-          maximumAge: 0, // 캐시 X
-          timeout: 5000, // 타임아웃 5초
-        }
+      const cleanup = myLocationTracking(
+        map,
+        setCurrentLatLng,
+        markerRef,
+        isTrackingRef
       );
-
-      // 언마운트 시 추적 종료
-      return () => {
-        navigator.geolocation.clearWatch(watchId);
-      };
+      return cleanup;
     } else {
       alert("이 브라우저는 위치 정보를 지원하지 않습니다.");
     }
   }, []);
+
+  // 마커 렌더링. 마커 배열이 변경될 때만 실행되도록
+  useEffect(() => {
+    if (mapInstance.current) {
+      renderMarkers(mapInstance.current, markers);
+    }
+  }, [markers]);
 
   // 추적 모드 활성화 시 현재 위치 중심으로 지도 이동
   useEffect(() => {
